refactor(core): move AgentBackend port contract to @maka/core/backend-types
Relocate the AgentBackend interface plus BackendCompactHistoryInput/Result
from the concrete ai-sdk-backend.ts implementation into the pure-type
backend-types module in core, where BackendSendInput/PermissionDecision
already live. Repoint the six runtime consumers (pi-agent-backend,
fake-backend, agent-run, ai-sdk-flow, session-manager, runtime-kernel) to
import the port from core instead of pulling in the ~2900-line backend
implementation for one interface. ai-sdk-backend re-exports the types for
backward compatibility. Type-only change; no runtime behavior change.
