+++ packages/core/src/backend-types.ts
@@ -3,16 +3,18 @@
  *
  * Source: V0.1_TECH_SPEC.md §13 + §6.1
  *
- * The full `AgentBackend` interface lives in @maka/runtime since
- * the implementations (AiSdkBackend / FakeBackend) are runtime
- * concerns. Core only exports the request/response shapes that cross the
- * runtime boundary.
+ * The `AgentBackend` port interface and the request/response shapes that
+ * cross the runtime boundary live here in @maka/core so that every backend
+ * implementation (AiSdkBackend / PiAgentBackend / FakeBackend) and their
+ * consumers depend on a small pure-type module, not on a concrete backend
+ * implementation file.
  */
 
-import type { AttachmentRef } from './events.js';
+import type { AttachmentRef, SessionEvent } from './events.js';
 import type { RuntimeEvent } from './runtime-event.js';
-import type { StoredMessage } from './session.js';
+import type { StoredMessage, BackendKind } from './session.js';
 import type { PermissionResponse } from './permission.js';
+import type { ContextBudgetDiagnostic } from './usage-stats/types.js';
 
 export interface BackendSendInput {
   /** AgentRun id for this invocation, when the caller has a run ledger. */
@@ -38,3 +40,22 @@ export interface BackendSendInput {
 
 /** Alias for clarity at the backend boundary. */
 export type PermissionDecision = PermissionResponse;
+
+export interface BackendCompactHistoryInput {
+  turnId: string;
+  runtimeContext: readonly RuntimeEvent[];
+}
+
+export interface BackendCompactHistoryResult {
+  contextBudget?: ContextBudgetDiagnostic;
+}
+
+export interface AgentBackend {
+  readonly kind: BackendKind;
+  readonly sessionId: string;
+  send(input: BackendSendInput): AsyncIterable<SessionEvent>;
+  compactHistory?(input: BackendCompactHistoryInput): Promise<BackendCompactHistoryResult>;
+  stop(reason: 'user_stop' | 'redirect'): Promise<void>;
+  respondToPermission(decision: PermissionDecision): Promise<void>;
+  dispose(): Promise<void>;
+}
